--- a/.env.example
+++ b/.env.example
@@ -1,2 +1,4 @@
 DB_USERNAME=postgres
 DB_PASSWORD=postgres
+EAGER_LOAD_DEV=true # only turn it on  when stress testing
+CI=TRUE # mark this true for benchmark testing of transactions api