--- a/app/controllers/api_docs_controller.rb
+++ b/app/controllers/api_docs_controller.rb
@@ -1,7 +1,7 @@
 class ApiDocsController < ActionController::Base
   include Swagger::Blocks
 
-  openapi_endpoint do
+  swagger_root do
     key :swagger, "2.0"
     info do
       key :version, "1.0.0"