This PR introduces performance-focused improvements to both our documentation and test infrastructure to help developers track and understand how the system behaves under stress and at scale.

📘 README Updates
Added a new "Performance" section to README.md that links to detailed performance documentation.

This section outlines:
   - System performance characteristics
   - Benchmarks and test reproducibility details

🧩 This gives developers a deeper technical understanding of how the app sustains efficiency under load.

📊 Benchmarking Infrastructure for Stress Tests
- Implemented a system to persist and analyze benchmark results from transaction stress tests:

✅ Features:
Saves benchmark results to:
 - tmp/benchmarks/transaction_benchmarks.csv

Captures:
      - Timestamp
      - Test name
      - Total execution time
      - Number of transactions
      - Average time per transaction
      - Automatically compares current test results to previous runs and surfaces deltas.

🔍 Run  trend analyzer via:

```bash
ruby spec/benchmark_analysis.rb
```

🧪 Race Condition Testing
  - Introduced controlled concurrent request simulations to identify potential race conditions under load.
  - This helps validate thread safety and correct behavior in high-throughput environments.

🧪 How to Test
      -  Run `bundle exec rake stress_test:transactions` to see the stress test on transactions.

## Stress test report
```bash
========================================================
Starting Transaction API Stress Test
========================================================
Configuration:
  - Number of users: 1000
  - Transactions per user: 50
  - Concurrency level: 20
  - Total transactions: 50000
========================================================

Creating 1000 test users...
....................................................................................................
Users created.
Generating 50000 transaction requests...
Processing transactions with 20 concurrent threads...
Progress: 100% complete (50000 succeeded, 0 failed)

========================================================
Stress Test Results
========================================================
Total time: 524.27 seconds
Total transactions: 50000
Successful transactions: 50000
Failed transactions: 0
Success rate: 100.0%
Transactions per second: 95.37
Average response time: 0.2096 seconds
Minimum response time: 0.1127 seconds
Maximum response time: 0.3145 seconds
========================================================

Verifying database state...
All database checks passed!
```